Move default_arguments to argument package.
