+++ rasa/cli/data.py
@@ -3,7 +3,7 @@ from typing import List
 
 from rasa import data
 from rasa.cli.arguments import convert as arguments
-from rasa.cli.default_arguments import add_nlu_data_param
+from rasa.cli.arguments.default_arguments import add_nlu_data_param
 from rasa.cli.utils import get_validated_path
 from rasa.constants import DEFAULT_DATA_PATH
