+++ rasa/cli/run.py
@@ -3,7 +3,7 @@ import logging
 import os
 from typing import List
 
-from rasa.cli.default_arguments import add_model_param
+from rasa.cli.arguments.default_arguments import add_model_param
 from rasa.cli.utils import get_validated_path
 from rasa.constants import (
     DEFAULT_ACTIONS_PATH,
